Adds basic DOM tree generation module
This adds a basic generator for random DOM trees. This generation
is reproducible so it can be used to benchmark by repeatedly
producing the same DOM structure.

Transparent elements are only able to host text nodes - it's not
worth programming in the complexity of transparent nodes just for
this generation.

Attributes are next on the list. After that, we will enrich the
generator by adding params.