Adjusts format of Element type, secures generators with params
This adds params to the generators in a way that makes them
accessible inside of any generator, using a `ReaderT`. `IORef` is
used to track the total nodes that have been constructed so we can
honor a maximum total nodes parameter to the generator. `Element`
is now a record, with a slightly richer type for handling child
elements.