Makes incremental improvements to underflow problem
Depending on the size of the document we try to generate, we are
experiencing an arithmetic underflow problem. We did some minor
shielding during element generation that prevents resources from
being depleted when counting total nodes, but no such protection
has been added for tracking max depth. I think the best approach
will be to remove the monad-based tracking of current depth and
replace it with a ride-along argument to all generators.