Adds structure generation
Rather than the weird stateful approach to generating, we're
instead generating a simple tree structure to represent conceptual
HTML - `Branch` is any element that can have children, `Leaf` is
any element that can only have text, and `Void` is a void element.

Next, we will convert from this generated skeleton type to the
`Element` types previously introduced, using the intersection of
sets for valid children and elements that correspond to the node
types defined above.