Adds breadth-first DOM generator to benchmarking suite
This DOM generator is breadth-first (populates a layer before
descending further into the tree), mostly spec-compliant (at least
enough that it should be able to convert to Brigid), respects user
parameters given via the `GeneratorParams` type, and doesn't
produce errors, arithmetic underflows, or hangs on DOM trees as
large as 100,000 total nodes.

This is done by tracking a collection of `Map`s that represent the
following:

- `layerMap`: A mapping of the current depth to UUIDs representing
nodes that are on that layer.
- `childMap`: A mapping of a UUID representing a node to a list of
UUIDs representing nodes that are children of that node.
- `nodeInfo`: A mapping of a UUID representing a node to details
about that node - namely, what element it is, and what node type
(`Branch`, `Leaf`, or `Void`) it is.