Treats certain terminal branch elements as leaf elements
During generation, if we had run out of total node or depth budget
prior to generating a branch element, we would just leave it empty.
Now, if it can be treated as a leaf element (that is, it can hold
raw text as valid content), we generate text to populate it.

This also updates the `translate` attribute generator to use the
new `YesNo` type.