Fixes safe base module
This is a lesson to make test cases in the same commit as the safe
module's creation.

Basically, the `ValidHref href Base` constraint couldn't be
reconciled when only the target is used. This issue would persist
in other like approaches, so the only option was to create separate
functions for each of the three possible cases. It's ugly, but it
works.