Adds support for the autocorrect global attribute
This is unfortunate because it's technically not global - it's
valid on a small handful of text input fields, and anything with
the `contenteditable` attribute. But because `contenteditable` is
a true global attribute, that makes this attribute technically
global. A safe module might be added in the future to ensure that
some `autocorrect` value is paired with `contenteditable`.