Fixes some attribute types, generators, and validity
This makes adjustments to valid tags and values for some
attributes (both in the various type families and in the element
generation). This involved making some changes to certain types to
support new permitted values that were previously unknown, as well
as changes to generators for some types and attributes.