Adds children to generated DOM to Brigid conversion
There are a couple blind spots still present in this, but for the
most part this has been implemented. It's unfortunate that it has
to be a large and tedious module casing on all valid children on a
per-element basis, but we can't make easy functions to return flow
and phrasing content because the compiler can't verify the elements
in those lists will be valid children for whoever uses them.

The next step will be to clean up the remaining TODOs. After that,
every function will be changed to return `Either String a` and
traversed so we can test that it will generate valid options for
everything and correct any errors that pop up.