Produces errors on incorrect DOM generation assignment
This leverages the `validation` package to collect all instances
where an incorrect scoped attribute or child is assigned to an
element, or when an attribute attempts to be built as a global
attribute when it is not. This will allow for us to create a
property test suite that can ensure we aren't generating invalid
attributes and children for any given element.