Adds test to ensure attribute generation, corrects ratio generator
This adds a test suite that, currently, tests each of the attribute
generators to ensure they work properly without throwing an
exception. We were dealing with arithmetic underflow exceptions
when generating some elements, but it wasn't until this test was
implemented that we were able to pin it down to the `imagesizes`,
`media`, and `size` generators. After additional experimentation,
it was discovered that the `ratio` generator having a `0` minbound
on the numerator generator was the issue - after increasing it to
`1`, none of these offending generators are throwing.