Adds tests for elements after making fixes to generation
This adds testing to ensure spec compliance (no validation failures
in the Brigid conversion step) for DOM generation. To ensure this,
a number of modifications to element and attribute validity were
made. Currently, a handful of known incomplete/broken elements are
removed from the testing, and will be added back in as they are
addressed.

A small note: recursive calls to `buildLayer` in the DOM generation
will cause the generator to scale down slightly (using `Gen.small`)
to avoid the size increasing out of control.