Adds element test cases
This tests each element type in turn to ensure that, for a depth
of 3 (the element, its children, and its grandchildren), no errors
occur in generation. Passing the test confirms that the element is
generated in a spec-compliant manner (at least as far as Brigid is
concerned).

Alterations were made to a number of elements to ensure that they
generate correctly so that they can pass. A few type synonyms were
introduced as well.