Makes Id and Name non-empty, adds Monoid instance for Class
`Id` and `Name` should never reasonably be empty, so we switched
their underlying types from `Text` to `NonEmptyText`, but kept the
to-text functions as additional helpers to minimize impact.

`Class`, however, can reasonably be empty, so we've given it a
`Monoid` instance. This changes the implementation of the `classes`
helper function, but not the actual effect.