Adds valid elements for `hx-validate`, stops building entities
`hx-validate` is valid on `input`, `select`, and `textarea` tags,
so these were added to the valid elements for this attribute.

Additionally, we no longer build entities when running the build
script. This should be done explicitly with the intent of updating
that type, not just wantonly so that changes get mixed into other
unrelated commits.