Adds support for event attributes
Currently, this is a lightweight approach of having one `on`
combinator that takes an `Event` ADT and the `RawJavaScript` it is
intended to run. Event attributes are technically global, but some
have no functionality on some elements. Since they can be used for
atypical means like as selector mechanisms or as poor man's
directives, we're not enforcing any requirements on which elements
they can appear. We may do this later with explicit combinators for
each event attribute, but there isn't any particular motivation to
do that at this time.