Widens valid children list for Paragraph tags
After further research, it was discovered that `p` tags can accept
the following fields under certain circumstances: `a`, `area`,
`del`, `ins`, `link`, `map`, and `meta`. The circumstances for when
these are valid children of `p` are either already covered by the
`ValidChild` mechanism (transparent tags) or cannot be checked for
(requiring a specific attribute, descendency at any depth), so we
allow them to maintain flexibility and trust the user to make the
right calls. This is better than being inflexible or forcing users
to use `customHTML`.